refactor: decompose the persona settings page
- +page.svelte drops from 2,278 lines to 280: state and handlers move
  into a co-located createPersonaPageState() rune module, the template
  splits into AppModeSection, AvatarGallery, CorePersonality,
  AiServicesSection, and StatsPanel, and the scoped styles move with
  the markup that owns them
- All effects stay in the page component; children receive the state
  instance via props
- No behavior change; one scoped selector became :global to keep
  applying to child-component roots (same runtime result)
